requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.5
django-crispy-forms==1.11.0
Markdown==3.3.3
mistune==0.8.4
Pygments==2.7.4
pytz==2020.5
sqlparse==0.4.1
wincertstore==0.2
